rename visualization_panels for clarity
