--- components/visualization_panel/README.MD
@@ -1,5 +1,0 @@
-The visualization panel contains the graphical representation of the modules and their interconnections. 
-
-Currently, this is implemented in [cytoscape](https://dash.plotly.com/cytoscape).
-
-When more visualization options are created, this component will be expanded.
