+++ components/visualization_panels/README.MD
@@ -1,0 +1,5 @@
+The visualization panels contain the graphical representations of the modules and their interconnections. 
+
+These visualizations are implemented in [cytoscape](https://dash.plotly.com/cytoscape).
+
+Right now there are different visualizations, one combining the active node, the user's pathway, and their search results, as well as a visualization focused on each of those individually.
