+++ components/visualization_panels/active_module_visualization.py
