+++ components/visualization_panels/search_results_visualization.py
